PR10280: re-enable 'force vermagic for guru-mode scripts'
Current version checking based on kernel ABI and build-ids is not
strict enough to prevent launching a stap module on a kernel version
it wasn't compiled for. This has the potential to crash a running
kernel, since ABI compatibility may not give a sufficient guarantee of
real compatibility.

This patch was reported to cause problems on a PPC machine, but the
reasons are still unclear. Worksforme with the PPC machine I tested it on,
but there are kernel patches that suggest the vermagic string may be
additionally mangled (43e24e83f35).

Re-enabling this and will cautiously monitor the aftermath.