Force correct order of evaluation of macro arguments in check_*register macros
Noted that a number of tests were failing on x86 machines with errors
like the following:

ERROR: register access fault [man error::fault] near identifier 'module_name' at
 /usr/share/systemtap/tapset/linux/context.stp:392:10

The problem was traced to the maxregno argument for the macro having a
?: operator which has lower precedence than || or >.  This caused the
conditional tests in check_fetch_register and check_store_register for
error reporting to incorrectly trigger.  Used ()'s in the conditionals
to force the correct order of evaluation.