@count() and @sum() should provide 0 for empty entries
When compiled with the bpf backend mmfilepage.stp would have a
segmentation fault when printing out recorded information in the probe
end.  This was cause by a foreach using an index from one global array
for other global arrays that did not have entries for some of those
indices.  The code computing the @count() and @sum() values did not
handle the situation where there were no statistics for that
particular index in the global array.  The proper default is that
@count() and @sum() in those cases should be 0.