PR24543 just-in-case fix :: stapbpf breaks when cpu0 is disabled
Certain perf_events constructs created by stapbpf must be pinned to
one cpu, it doesn't matter which. Previously cpu0 was used.

*Very rarely* it's possible for cpu0 to be disabled.
(It's not recommended,

In that case stapbpf will fail with a super confusing error.

This patch improves readability and allows stapbpf to use a fallback cpu.
cpu0 is still preferred whenever available.

* stapbpf/stapbpf.cxx (default_cpu): New global.
  (mark_active_cpus): Set default_cpu, prefer cpu0 whenever possible.
  (create_group_fds): Use default_cpu, not cpu0.
  (register_uprobes): Ditto.
  (register_kprobes): Ditto.
  (register_tracepoints): Ditto.
  (register_timers): Ditto.
  (register_perf): Ditto.
  (load_bpf_file): Clarifying comment on when active cpu info is checked,
    use default_cpu for uctx bpf_transport_context.