PR23285: stapbpf/procfs: use effective-userid consistently
The euid is easy and consistent in the construction of the
stapbpf/procfs path name.  Use the umask() field at least for
the fifo nodes (should also on the parent directory probably).
For the stapbpf -v case, print some fifo-related diagnostics.
Split out the .stp test case into a separate file, for easier
hand-execution.