PR22315: error handling support for stapbpf
This patch introduces enhancements for the BPF backend: try-catch block support,
error probes and an improved error tapset that works with the try-catch block
implementation.

The error tapset works as follows: if the error is called from a try block, the
program's execution will jump to the corresponding catch block. If the error is
called outside a try block, it gets registered and the program will terminate.

Epilogues are now added to each probe. The epilogue checks if an error occurred,
and calls a perf_event to print the error message. It also performs a check to
see if a MAXERRORS constraint is violated and if so, it signals a hard error to
the runtime via another perf_event.

It's worthwhile noting that error messages are copied into userspace (when the
error is registered) and then printed later (during epilogue). This is because
the BPF stack space is very limited which means the string cannot be efficiently
stored on the stack. If future versions of BPF permit more stack space, the
error message string could be stored directly onto the BPF stack instead of
storing it in userspace.

1) Added try-catch block generation and error probe handling.
2) Updated error tapset to work with try-catch statements.
3) Added new internal global variable, perf_events and BPF assembly statements.
4) Updated prologue to include a check for hard errors.
5) Designed epilogue which prints error messages and signals hard errors.
6) Added mapping for resource constraints set with -D.
7) Added test cases for the error tapset and try-catch blocks.
8) Updated NEWS.