translate.cxx: Make stap-symbols.h a separate CU.
This increases the kernel module compilation parallelism level.

The Pass-4 is now about 700ms faster when --ldd is used for a
simple .stp script file using print_ubacktrace() on an Intel
Core i9-9900K machine.

For the dyninst runtime, the new CU is still compiled into the
resulting .so file but this backend is not using the symbols yet
and the compilation is not parallelized.

Thanks fche for the review.