PR25290: process(EXE).begin probes might not get fired.
Occasionally the utrace callbacks might get called in the session state
of STAP_SESSION_STARTING so we have to be prepared to that otherwise
the user probe handler would never be called.

Also added a lot of debugging log output controlled by -DDEBUG_PROBES
and -DDEBUG_TASK_FINDER=LEVEL.

Thanks fche for the help.