The "Build-id mismatch" error now becomes a warning
It is normal for different binaries with the same file path coexist in
the same system via chroot or namespaces, therefore we make sure
build-id mismatch is really a warning instead of an error which leads
to non-zero exit codes of staprun.

For safety, we still make sure we do not register uprobes and etc to
wrong binaries when the build-id check fails.