RHBZ1788662: blacklist kernel cpu_idle tracepoint
This tracepoint is problematic because it is called from an idle
context, where RCU/lockdep is not legal to call.  On lockdep kernels,
RCU warnings and even possibly-related panics have been reported.

Kernel tracepoint handlers protect themselves by wrapping their
innards in rcu_irq_enter/rcu_irq_exit(), which flips the legality flag
back on (even during idle), but these functions are not
module-exported, and it's not clear they'd be sufficient anyway.  So
we blacklist in non-guru mode.