RHBZ1795196: tolerate partial elf + missing dwarf vmlinuz for -d kernel
Previous code in dump_symbol_tables() couldn't tolerate the case where
new elfutils opened /boot/vmlinuz* as an elf file for the kernel, and
could not extract a symbol table from that (nor an absent -debuginfo).
New code instead emits a warning and moves on.  A special error code
triggers return to the dump_kallsyms() path that prior elfutils/stap
versions rely on.