RHBZ1810216: fix ftrace() tapset function related autoconf test
It was reported that the ftrace() tapset function stopped working at
some point - due to an autoconf-* false negative.  We failed to notice
because the tapset function was quietly mapped to a nop in this case.
Now ftrace() will #error in -p4 if autoconf doesn't detect a suitable
kernel facility.  A forked testcase buildok/logging-embedded2.stp
checks for it.

Suggested-by: Oleksandr Natalenko <onatalen@redhat.com>