PR25686: Fix ambiguous octal escapes in _stp_text_str
Tapset functions like user_buffer_quoted that use the _stp_text_str function
for octal escaping would produce incorrect output if a null byte was followed
by a byte that represents a printable digit in US-ASCII. The null byte would
be emitted in the shorthand \0 form, then the printable digit would be emitted
literally, resulting in (e.g.) \06 instead of \0006 for the two input bytes
"0x00 0x06".

Fix by disabling shorthand escapes for null bytes. The alternatives are to:

(a) look ahead in the buffer to see if the next byte is a digit and only
expand the \0 to \000 if it would be ambiguous; or
(b) octal-escape all printable digits

The second would be woeful for the intended function of human-readable-ish
buffer representations. The former seems potentially fragile and risky,
so I just went with simple.