Correct error counting error in systemtap_session::print_error()
Logic such as

  if (verbose > 0 || seen_errors[pe.errsrc_chain()]++ < 1)

short-cirtuits the increment of seen_errors[] in verbose mode.
This can lead systemtap_session::num_errors() to return 0 (!!)
even if errors occurred, which can in turn lead to later passes
being attempted.  This sort of thing is nuked in two places,
to look thusly instead:

  if (verbose > 0 || seen_errors[pe.errsrc_chain()] < 1)
    {
      seen_errors[pe.errsrc_chain()]++;
      ....
    }

The above problematic pattern was introduced by some guy in commit
0e1d5b7eb39785cb.  He should never again be allowed near computers.