translator: ditch one assert(0)
In translator_output::assert_0_indent(), we used to assert(0) if
something went wrong with the translation process that resulted in the
generation of incorrectly nested C code.  We could emit a warning or a
runtime_error or something, but even doing nothing is OK, since a
pass-4 error will surely follow.