Avoid exceeding space constraints for BPF environment in mmwriteback.stp
BPF is very contrained on the space allowed when running the code.
Using too complicated printf statements caused too much stack space to
be used for temporaries and the stap compile would fail for the BPF
backend.  Replacing the complicated printf with simpler string printf
statements uses less stack space and allows the BPF backend to compile
this example.