Make dtrace generated code work with LTO (take 2)
LTO needs to know which variables might be accessed by code.  In sdt.h
there is a fair amount of assembly code which LTO cannot analyze.  As
a result LTO would assume that the semaphores variables for various
userspace probes were not accessed by the generated code.  The linking
would fail because LTO would optimize away the semphore variables
causing undefined references.

The fix adds the semaphore variables as input operands to the assembly
statements.  This gives the LTO analyzer information that the
variables are used within the assembly statements and should not be
removed.