task_finder2: fix memory leak when task workers fail to get added
None of the error paths for the __stp_tf_task_work_add() calls free the
tf_work allocation when the task_work_add fails. This fixes that.

This also makes a nitpick to __stp_tf_task_worker_fn() to reduce the
critical section of __stp_tf_task_work_list_lock.

Reported-by: Frank Ch. Eigler <fche@redhat.com>

Signed-off-by: Yichun Zhang (agentzh) <yichun@openresty.com>