Add CONFIG_COMPAT 32-bit support for aarch64 and powerpc
The aarch64 and powerpc kernels may be built with support for 32-bit
user-space applications.  On Fedora 33 the aarch64 kernel is
configured with CONFIG_COMPAT enabled.  Needed to provide some support
in _stp_is_compat_task2 for the aarch64.  The Fedora 33 ppc64le
kernels by default do not have CONFIG_COMPAT enabled, but included
similar check in there for powerpc in case locally built kernels
enable it.