PR27802: tweak configure.ac for AC_SUBST(pkglibdir)
autoconf is such a hassle for getting expanded path names for the
--FOOdir=BAR names.  During execution of the configure script, they
may be defined to be expressions with literal $values in them, which
need to be eval'd with other $values preset.  What a mess.  Tweak the
sequence used to predict the $pkglibexecdir through heuristics to get
around all this indirection.  (e.g. /usr/libexec/systemtap), whether
or not --libexecdir= is given.