Further tweak deviceseeks.stp example to work on Fedora 32/33/34
The previous fix for Fedora rawhide Linux kernels (5.12) did not work
on earlier Linux 5.11 kernels in Fedora 32/33/34 because struct bio
has a different field, bd_disk that has a pointer to the queue.
Separated the process for getting the sector size into two steps:

-find the request queue struct
-find the sector size for the device

This script has been tested on RHEL8, Fedora 34, and Fedora Rawhide.